+++ src/pages/CourseList/hooks/usePagination.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { COURSES_PER_PAGE } from '../../../constants/pagination';
 import { ICourse } from '../../../interfaces/Course.interface';
@@ -6,9 +7,11 @@ const usePagination = (courses: Array<ICourse>) => {
   let [searchParams, setSearchParams] = useSearchParams();
   let currentPage = parseInt(searchParams.get('page') ?? '1');
 
-  const idxOfLastCourse = currentPage * COURSES_PER_PAGE;
-  const idxOfFirstCourse = idxOfLastCourse - COURSES_PER_PAGE;
-  const currentCourses = courses.slice(idxOfFirstCourse, idxOfLastCourse);
+  const currentCourses = useMemo(() => {
+    const idxOfLastCourse = currentPage * COURSES_PER_PAGE;
+    const idxOfFirstCourse = idxOfLastCourse - COURSES_PER_PAGE;
+    return courses.slice(idxOfFirstCourse, idxOfLastCourse);
+  }, [courses, currentPage]);
 
   const paginate = (number: number) => {
     setSearchParams(`page=${number}`);
